Refactor configuration and domain helpers into package
